retry to upload after error

+++ backup.sh
@@ -79,18 +79,20 @@ LOCAL_FS="$SCRIPT_DIR/fs.zip"
 LOCAL_DB="$SCRIPT_DIR/db.zip"
 
 # FILESYSTEM
-echo "~~~ Zip filesystem"
-cd "$PROJECT_PATH" && zip -r "$SCRIPT_DIR/fs.zip" . \
-	-x ".git/*" \
-	-x "storage/framework/cache/*" \
-	-x "storage/framework/sessions/*" \
-	-x "storage/framework/debugbar/*" \
-	-x "storage/framework/views/*" \
-	-x "storage/logs/*"
+if [ ! -f "$LOCAL_FS" ]; then
+  echo "~~~ Zip filesystem"
+  cd "$PROJECT_PATH" && zip -r "$SCRIPT_DIR/fs.zip" . \
+    -x ".git/*" \
+    -x "storage/framework/cache/*" \
+    -x "storage/framework/sessions/*" \
+    -x "storage/framework/debugbar/*" \
+    -x "storage/framework/views/*" \
+    -x "storage/logs/*"
 
-if [ -n "$SSL_PATH" ]; then
-  echo "Zip SSL certificates"
-  zip -rg "$SCRIPT_DIR/fs.zip" "$SSL_PATH"
+  if [ -n "$SSL_PATH" ]; then
+    echo "Zip SSL certificates"
+    zip -rg "$SCRIPT_DIR/fs.zip" "$SSL_PATH"
+  fi
 fi
 
 echo "Uploading $DEST_FS_KEY to s3"
@@ -103,46 +105,48 @@ else
 fi
 
 # DATABASES
-echo "~~~ Dump databases"
-if [ -n "$DB_LIST" ]; then
-  if ls "$SCRIPT_DIR"/*.sql 1> /dev/null 2>&1; then
-    rm "$SCRIPT_DIR"/*.sql
-  fi
-
-  IFS=',' read -ra DBS <<< "$DB_LIST"
-
-  if [ ${#DBS[@]} -eq 0 ]; then
-  	echo "No valid database entries found in DB_LIST."
-  	exit
-  fi
-
-  TIMESTAMP=$(date +%Y-%m-%d_%H-%M)
-
-  for ENTRY in "${DBS[@]}"; do
-    [ -z "$ENTRY" ] && continue
-
-  	IFS=':' read -r DB_NAME DB_HOST DB_USER DB_PASS <<< "$ENTRY"
-
-  	echo "Dumping $DB_NAME..."
-  	mysqldump -h "$DB_HOST" -u "$DB_USER" -p"$DB_PASS" "$DB_NAME" 2>/dev/null > "${SCRIPT_DIR}/${DB_NAME}_${TIMESTAMP}.sql"
-
-    if [ "${PIPESTATUS[0]}" -eq 0 ]; then
-      echo "Successfully dumped $DB_NAME"
-    else
-      echo "Error: Failed to dump $DB_NAME"
+if [ ! -f "$LOCAL_DB" ]; then
+  echo "~~~ Dump databases"
+  if [ -n "$DB_LIST" ]; then
+    if ls "$SCRIPT_DIR"/*.sql 1> /dev/null 2>&1; then
+      rm "$SCRIPT_DIR"/*.sql
     fi
-  done
 
-  if ls "$SCRIPT_DIR"/*.sql 1> /dev/null 2>&1; then
+    IFS=',' read -ra DBS <<< "$DB_LIST"
+
+    if [ ${#DBS[@]} -eq 0 ]; then
+      echo "No valid database entries found in DB_LIST."
+      exit
+    fi
+
     TIMESTAMP=$(date +%Y-%m-%d_%H-%M)
 
-  	DB_ZIP="${SCRIPT_DIR}/db.zip"
-  	zip -j "$DB_ZIP" "$SCRIPT_DIR"/*.sql
-    rm "$SCRIPT_DIR"/*.sql
+    for ENTRY in "${DBS[@]}"; do
+      [ -z "$ENTRY" ] && continue
 
-  	echo "Combined archive created: $DB_ZIP"
-  else
-  	echo "No SQL files found to archive."
+      IFS=':' read -r DB_NAME DB_HOST DB_USER DB_PASS <<< "$ENTRY"
+
+      echo "Dumping $DB_NAME..."
+      mysqldump -h "$DB_HOST" -u "$DB_USER" -p"$DB_PASS" "$DB_NAME" 2>/dev/null > "${SCRIPT_DIR}/${DB_NAME}_${TIMESTAMP}.sql"
+
+      if [ "${PIPESTATUS[0]}" -eq 0 ]; then
+        echo "Successfully dumped $DB_NAME"
+      else
+        echo "Error: Failed to dump $DB_NAME"
+      fi
+    done
+
+    if ls "$SCRIPT_DIR"/*.sql 1> /dev/null 2>&1; then
+      TIMESTAMP=$(date +%Y-%m-%d_%H-%M)
+
+      DB_ZIP="${SCRIPT_DIR}/db.zip"
+      zip -j "$DB_ZIP" "$SCRIPT_DIR"/*.sql
+      rm "$SCRIPT_DIR"/*.sql
+
+      echo "Combined archive created: $DB_ZIP"
+    else
+      echo "No SQL files found to archive."
+    fi
   fi
 fi
